Change default value for the priority attribute
The default value for the priority attribute is incorrectly being set to "inherit". That is not a possible value for the priority attribute of a sitemap.xml file.  Google bot throws an error on it. This only seems to be an issue when a user that does not have rights to the Extended Attributes tab creates a page. Because they have no rights to the tab the default value is being used.  The default value is not used when users have rights to the tab. Instead it uses the first available value in the drop down which is "0.1".